Updated to final badge for the plugins package
Azure DevOps Artifacts doesn't allow a badge to be generated for a package until it has a release version. As a result, for the first version, the badge can't be assigned until _after_ the first version has been published to the package repository. As such, since this is an otherwise new package, we need to loop back to update the `README` to include the correct badge. (Previously, this used the badge for the `OnTopic.Editor.AspNetCore` project.)