Bug Fix: Add server-side validation for correct `Key`
Previously, for a new topic, an `InvalidKeyException` was displayed to the interface. For an existing topic, however, invalid characters were removed, and the key was renamed. Neither are ideal. Instead, this returns a friendly error message whenever an invalid key name is proposed. This resolves #11 and the server-side aspect of #13. (Client-side handling can be done via data configuration.)