Bug Fix: Redirect on Rename
Currently, there's a redirect that occurs when a topic is new, to ensure that it ends up on the newly created page. But there's not a similar redirect if the page has been renamed, and thus the current URL is no longer valid. Better yet, we can _always_ redirect; that will solve both of the above issues, while also helping prevent accidental form resubmission on refresh. This resolves #12.