Improved styling of server-side modeling errors
Server-side modeling errors are injected into the `asp-validation-summary` as a bullet list. For consistency, I've wrapped this in a Bootstrap `alert` panel, and reformatted the bullet list so that the margins are more consistent with other Bootstrap panels. This required special handling of the `ul` and `li` since those are effectively eliminated by the ExtJS CSS.