Update precision of `SetVersion()` calls to match `Version`
In OnTopic 5.0.0, we updated the `VersionHistory` to use `DateTime` instead of `SqlDateTime`, and updated the underlying SQL database to use `DATETIME2(7)` instead of `DATETIME`, thus allowing both OnTopic and SQL to use the same precision for version numbers. In doing so, however, the OnTopic Editor's _Versions_ dropdown box was never updated to use the higher-precision format. This resolves this bug, as closes out Issue #18.