Ensure redirect to grandparent when deleting nested topics
Normally, when a topic is deleted, the editor is redirected to the parent topic. In the case of a nested topic, the parent topic is a `List` topic, which is an infrastructure element not intended to be edited directly. Instead, nested topics should be redirected to the grandparent upon deletion. The logic for this is already in place, but doesn't work since the code is trying to evaluate the `ContentType` from `parent.Attributes`, not `parent.ContentType`. This addresses #36.