Ensured that view models from `OnTopic.ViewModels` are loaded
Currently, some areas of the OnTopic Editor rely on view models that are defined in the `OnTopic.ViewModels` library. For example, the `QueryableAttributeDescriptor`'s `RootTopic` maps to a `TopicViewModel`, and expects that the target type's `ContentType` will map to a view model which derives from `TopicViewModel`. So if it's pointing to a metadata lookup list, it will expect a `ListTopicViewModel`.

This will be resolved in #29 with support for `[MapAs()]`. Until then, however, the `ITypeLookupService` must include the `TopicViewModelLookupService`. We may be able to revisit this once #29 is in place, as then we can look for a simpler alternative to relying on `TopicViewModel` as a base topic, which will make the OnTopic Editor independent of the `OnTopic.ViewModels` library.