Increase `ResultLimit` for `TopicList` view component
The `TopicListAttributeDescriptorViewModel` doesn't currently define a `ResultLimit`, and instead hard-codes a value in the `TopicListViewComponent`. We'll likely revisit this later, to provide more flexibility in configuration. Until then, however, we're increasing the `ResultLimit` from `100` to `250`. This addresses an issue with, in particular, binding to the `Country` metadata; since there are over `100` countries, this list was getting truncated, thus preventing some countries from being selected. This resolves #16.