Updated defaults for exporting topics
Nearly every time we export topics, we want to include child topics and nested topics. These are disabled by default in `ExportOptions`, by design, as we want this to be explicit when working with the OnTopic Data Transfer library via code. When exposed via the interface, however, we want to default to the expected use case, and especially since the options are clearly presented and can easily be opted out of.

To facilitate, the `ExportOptions` on the `ExportViewModel` are initialized to set `IncludeChildTopics` and `IncludeNestedTopics` to true. This resolves Issue #34.