Explicitly evaluate `Topic.Key` as part of `TopicQueryService`
Previously, the `TopicQueryService` exclusively searched the `Topic.Attributes` for matches. With the migration of `Key` from `Topic.Attributes` to being a first-class citizen of `Topic`, that means `Key` is no longer searched. This may not be a problem if a topic has a `Title` attribute, as that will usually overlap with the `Key`. But it's a problem if the topic doesn't include a `Title`. It can also be an issue if the search query matches the `Key` but not the `Title` (e.g., `CustomerService` won't match a `Title` of "Customer Service").

To remedy this, the `TopicQueryService`'s `IsValidTopic()` is extended to explicitly check the `Topic.Key` alongside the `Topic.Attributes. This isn't especially pretty, but it solves the problem. This resolves #15.