Introduced labels for `DefaultValue`, `ImplicitValue`
If the value is inherited, has a default value, or has an implicit value, then the first item in the select box should indicate what that is. The `FileList` already handled `InheritedValue`, but not `DefaultValue` or `ImplicitValue`. This extends that functionality.

We don't want to select the actual value in case it's implicit or inherited, as we want the source that's setting it to continue to control that. And, if it's a default, we don't want to set that until after the editor has saved the form, in case they decide to first set a `BaseTopic` (in which case, `DefaultValue` shouldn't be applied).

This addresses the `FileList` portion of #26.