Exposed both `CurrentValue` and `Value` to `LastModified(By)` attributes
Previously, the `CurrentValue` was displayed—which meant it was empty on a new form. Now, the `CurrentValue` is displayed conditionally if it exists, and the new `Value` is _always_ displayed, so that the editor can see what the new value will be.

In addition, I've introduced a fallback for `InheritedValue` and `ImplicitValue`, in order to satisfy the `LastModified` and `LastModifiedBy` expectations of #26. In practice, this is likely a rare scenario, and will only affect the `CurrentValue`.