Updated `LastModifiedBy` to allow configurable `DefaultValue`
Previously, the `DefaultValue` for `LastModifiedBy` was hard-coded to "System". By removing this, we allow this value to instead by set by the `DefaultValue` in the editor. This gives more flexibility and control in scenarios where the OnTopic Editor is not authenticated. That said, in most scenarios, the OnTopic Editor will be authenticated, so this won't come into play. This resolves #39.

This further satisfies the `LastModifiedBy` needs of #26 by also exposing `InheritedValue` and `ImplicitValue` in case neither `Value` nor `DefaultValue` are set.

Note: This currently won't work due to a bug in how `DefaultValue` is assigned (#25); this bug is expected to be fixed prior to 5.1.0 release.