Introduced fallback for `MaximumLength`
This might be better handled via a helper method on these classes, or even more intelligent code for the property getter which limits the `MaximumValue` based on the `IsExtendedAttribute` value. For now, however, this gets the job done. Once we update to OnTopic 5.1.0 (alpha) we can reevaluate if this makes more sense in the constructor, since that will allow constructor mapping. This satisfies the needs of #32.