Override Bootstrap's (new) `passing-left` for `ol`, `ul` lists
By default, Bootstrap seems to introduce a large 2rem (!) `padding-left` value for all `ol` and `ul` lists. That seems excessive, but that may get mitigated by other Bootstrap selectors that aren't applying to our markup? Notably, this interferes with the ExtJS `TreeView` controls, causing each level to be excessively indented. I dislike resetting such a fundamental elements, and we may want to revisit this later, but for now this resolves the issues with the trees, without introducing any obvious side effects.