Added `form-label` class to primary labels
In Bootstrap 5.0.0, each `label` needs an explicit `form-label` class. I'm unsure why, but it's easy enough for us to add. There may be labels within individual attributes that need to be addressed as well, but this at least addresses the primary attribute labels that are added via the `_Layout.cshtml` which each attribute is wrapped in. This resolves an issue, at minimum, where the bottom margin on the labels were reduced.