Removed unnecessary "close" buttons in alerts
The alerts in `Edit` are meant to be permanent fixtures related to the current state, not temporary notifications that can be dismissed. Even if they were dismissed, they would just end up appearing again the next time the page is refreshed. Given this, I've removed the "close" buttons.

(The format for these has changed in Bootstrap 5.0.0, thus requiring these be revisited. Since they're not needed, it makes more sense to remove them rather than upgrade them.)