Override Bootstrap's new default of underlining links
There's a good argument for underlying links, since it creates a clear indication that the text is, in fact, linked. In our case, all of our links in the editor are part of the navigation or lists which are very clearly linked. Overriding this globally.

As with the list padding (1b76a3a), I'm not a fan of resetting this globally, but this should be a safe update since it doesn't affect layout; if we find cases where it interferes with Bootstrap's assumptions, we can revisit limiting the scope of the selector later on.