Removed `row` class from `IsModal` window
It's not entirely clear why, but having the overarching `row` class around the `IsModal` window causes it to exceed its boundaries due to how Boostrap calculates the `padding-left` of the `row` class. This can easily be mitigated by removing `row` if `IsModal`.

This doesn't _appear_ to have any off-target effects, but I'm not entirely clear why a `row` was placed here in the first place, so this should be bounced off of @ktrunkey for validation.