Prevent inheritance of topic references for attributes
By default, both attributes and topic references transparently inherit from the `BaseTopic` if a) a `BaseTopic` exists, and b) the local value is `null`. This is the desired behavior in almost all cases.

The exception is when evaluating the values for display in the OnTopic Editor, where displaying inherited values results in copying the inherited value to the local topic on `Save()`. The call to `Topic.Attributes.GetValue()` already sets `inheritFromBase` to false, but the call to `Topic.References.GetValue()` didn't—likely because inheritance of topic references was a relatively late addition in OnTopic 5.0.0.

Regardless, this is fixed now. This resolves Issue #31.