Set hidden value for `IncomingRelationshipAttribute`
With the new validation logic, every attribute type plugin is expected to submit a `Key`, `ContentType`, and `Value` (70f9b5a). If not, the `EditorController` will suspect a problem with the model binding or attribute type plugin view. This normally isn't an issue. But the `IncomingRelationshipAttribute` is a special read-only attribute that's only intended to give reference information. As these types are an exception to begin with, this can be mitigated by adding a hidden field for `Value`, just as we already do in a similar situation for `InstructionAttribute`.