Ensure that `DefaultValue` is being set when appropriate
If the `BaseTopic` is not set and the attribute value is empty, attempt to fall back to the `DefaultValue`. This fixes a bug where the `DefaultValue` wasn't being correctly set (#25). This logic doesn't check to see if `DefaultValue` is defined; if it isn't, the value will be removed, if it already existed, which is the expected behavior.