Disabled `required` in attribute type views if `DefaultValue` is defined
All attribute labels will be marked as required if `IsRequired` is defined. If a `DefaultValue` is defined, however, then the value shouldn't _actually_ be required, as it will fallback to the `DefaultValue`. This logic is already handled on the server-side (00fa995); this commit extends it to the client-side. This is done with the help of a new helper property, `IsValueRequired`, on `AttributeDescriptorViewModel`, which combines `IsRequired` and `DefaultValue`. This satisfies the client-side portion of #25.