Update EditorController.cs
The original finding that radio buttons would an empty value if not selected was incorrect, and based on a faulty test case. Optional radio buttons that are not set _will_ be treated as `null`, not empty, as initially reported. As a result, the binding model validation logic is being updated to accept this potentiality.

This reduces its ability to detect faulty attribute type plugins since it can't tell the difference between e.g. a radio button and a view that is not correctly setting or binding the `Value` field. This will be up to attribute type plugin developers to test and prevent.

This applies to Issue #43.