Introduced new `CoreTopicViewModel`
The `CoreTopicViewModel` is intended to map the common properties from `Topic` that are expected to be broadly required within the OnTopic Editor. This includes not only shared properties required by both `EditingTopicViewModel` and `AttributeDescriptorViewModel`, but also the topic references, such as `BaseTopic` and `RootTopic`, required by the attribute descriptor view models, such as `TopicReferenceAttribute` and `IncomingRelationshipAttribute`. This will be the core class used to resolve #29.