Updated `QueryableTopicList` to use `CoreTopicViewModel`
Updated the `QueryableTopicListAttributeDescriptorViewModel` to use the new `CoreTopicViewModel` (17ac6f0) for the `RootTopic` property. This required updating it to use the new `[MapAs()]` feature from the `TopicMappingService` to ensure that it doesn't instead try to map it as e.g. `{ContentType}TopicViewModel`, which may not exist.

This is the primary use case identified by #29, though there are other instances we still need to address.