Updated `TopicReference` to use `CoreTopicViewModel`
Updated the `TopicReferenceAttributeDescriptorViewModel` to use the new `CoreTopicViewModel` (17ac6f0) for the `RootTopic` property. This required updating it to use the new `[MapAs()]` feature from the `TopicMappingService` to ensure that it doesn't instead try to map it as e.g. `{ContentType}TopicViewModel`, which may not exist.

This is very similar to the previous update to the `QueryableTopicListAttributeDescriptorViewModel`—and, in fact, once we implement #27, we'll be relying on that inheritance for `RootTopic`.

In the meanwhile, however, this satisfies the needs of #29.