Updated `IncomingRelationship` to use `CoreTopicViewModel`
Updated the `IncomingRelationshipAttributeDescriptorViewModel` to use the new `CoreTopicViewModel` (17ac6f0) for its `RelatedTopics` collection property. This required updating it to use the new `[MapAs()]` feature from the `TopicMappingService` to ensure that it doesn't instead try to map it as e.g. `{ContentType}TopicViewModel`, which may not exist.

This addresses the `IncomingRelationships` portion of #29.