Updated `EditingTopicVIewModel` to use `CoreTopicViewModel`
Updated the `EditingTopicViewModel` to use the new `CoreTopicViewModel` (17ac6f0) for its `BaseTopic` property. This required updating it to use the new `[MapAs()]` feature from the `TopicMappingService` to ensure that it doesn't instead try to map it as e.g. `{ContentType}TopicViewModel`, which may not exist.

This addresses part of the `EditingTopicViewModel` portion of #29. We still need to update it to derive from `CoreTopicViewModel` as well.