Updated `AttributeDescriptorViewModel` to derive from `CoreTopicViewModel`
The `AttributeDescriptorViewModel` previously derived from `TopicViewModel`, thus maintaining an otherwise unnecessary (and unwelcome) dependency on `OnTopic.ViewModels`. With the introduction of `CoreTopicViewModel` (17ac6f0), this can now be used as the base class, thus providing the same common properties that it depends on, without maintaining the dependency.

This satisfies the `AttributeDescriptorViewModel` requirements of #29.