Updated `EditingTopicViewModel` to derive from `CoreTopicViewModel`
The `EditingTopicViewModel` previously derived from `TopicViewModel`, thus maintaining an otherwise unnecessary (and unwelcome) dependency on `OnTopic.ViewModels`. With the introduction of `CoreTopicViewModel` (17ac6f0), this can now be used as the base class, thus providing many of the same common properties that it depends on, without maintaining the dependency.

That said, there are some properties that `EditingTopicViewModel` depends on that other implementers of `CoreTopicViewModel` don't need, such as `Parent` and `LastModified`. These need to be reimplemented here in order to maintain compatibility. Unlike the original `TopicViewModel`, the new `Parent` only needs to map the immediate parent, and doesn't need to crawl all ascendants.

This satisfies the `EditingTopicViewModel` requirements of #29.