Remove `TopicViewModelLookupService` from host project
Now that the OnTopic Editor project no longer depends on `OnTopic.ViewModels`, in preference for the local `CoreTopicViewModel` (17ac6f0), we can remove the dependency on `TopicViewModelLookupService`.

This was recently introduced to resolve the fact that e.g. `BaseTopic` and `RootTopic` references weren't being properly mapped if the content types they pointed to weren't in the OnTopic Editor, but were in the View Models library (5ee797cc).