Namespaced `data-*` attributes for Bootstrap dropdown component
In Bootstrap 5.0.0, the data attributes have been namespaced, so `data-toggle` is now `data-bs-toggle`. This resolves a bug where the dropdown stopped working, as noted in #41.

While I was at it, I updated the `aria` accessibility attributes and and markup to align with the reference standard from the Bootstrap documentation. (The markup is flexible, but it's easier to debug issues when the markup is aligned.) I also introduced a `nobr` to prevent wrapping of the dates.