Updated `data-*` namespaces for Bootstrap modal component
In Bootstrap 5.0.0, the data attributes have been namespaced, so `data-toggle` is now `data-bs-toggle`. This resolves a bug where the tabs stopped working, as noted in #41. This includes both the actual modal definition in `NestedTopicList`, as well as the close button in `_Toolbar`.

While I was at it, I added a missing `aria` attribute.

Unlike previous updates, this alone doesn't resolve the issues with the modal windows; they will require more attention.