Namespaced `data-*` attributes for nav/tab component in `Import`
In Bootstrap 5.0.0, the data attributes have been namespaced, so `data-toggle` is now `data-bs-toggle`. This resolves a bug where the tabs stopped working, as noted in #41.

While I was at it, I updated the `aria` accessibility attributes and and markup to align with the reference standard from the Bootstrap documentation. (The markup is flexible, but it's easier to debug issues when the markup is aligned.)

These are identical to the changes made previously to `_Toolbar` (258d0b1); as part of that, I had missed that we also had tabs on the `Import` page.