Added script for initializing the tooltips
In Bootstrap 5.0.0, the tooltips are no longer automatically initialized by Bootstrap. Instead, they need to be wired up individually via JavaScript. I've added a new script to `Layout` that does exactly this, by looping through each `data-bs-toggle="tooltip"` element and initializing it as a `bootstrap.Tooltip()`. This returns the tooltips to a functioning state, thus satisfying the tooltip line item in #41.