Updated `Modal` script to use new Bootstrap modal syntax
With Bootstrap 5.0.0, the components are no longer implemented via jQuery, and need to be initialized via the `bootstrap` object—e.g. `bootstrap.Modal()` for modal windows, or retrieving an instance of the object via `bootstrap.getInstance()`. While I was at it, for consistency, I updated the scripts to prefer straight JavaScript instead of awkwardly mixing jQuery and plain JavaScript. The jQuery made more sense previously when it was also used for the modal windows.

There was also some code in here that is no longer necessary, such as setting the first tab as active. That JavaScript is now removed.

This resolves the modal functionality in #41.