Moved logic for setting modal window height to `Modal` script
Instead of setting the height of any and all modal windows at the beginning of each page load, instead load it when the modal window is opened. This simplifies the code since we already e.g. have a reference to the editor frame being targeted. While I was at it, I migrated from jQuery to JavaScript syntax.