Set `AttributeBindingModelBinder` to fallback to `AttributeBindingModel`
The `ITypeLookupService` in OnTopic 5.0.0 now supports setting fallbacks, which will be used if a primary type cannot be identified. By setting the `AttributeBindingModelBinder` to fallback to `AttributeBindingModel`, we remove the _requirement_ for attributes to define their own `AttributeBindingModel` class; if they don't need to e.g. customize the `GetValue()` method, they can just rely on the standard out-of-the-box implementation.