Deleted empty derivatives of `AttributeBindingModel`
With the exception of the `BooleanAttributeBindingModel`, all of the rest of the derivatives of `AttributeBindingModel` added no logic, and were simply there to satisfy the preexisting need for each attribute type plugin to define one, just in case it was needed. This is ugly, and makes it harder to understand the plugin structure by adding additional levels of abstraction. With the fallback for `AttributeBindingModel` in place (412ee6e), this can be removed.