Derived `TopicReference` attribute descriptor from `TokenizedTopicList`
The `TopicReferenceAttributeDescriptorViewModel` shared the `RootTopic` with the `QueryableTopicListAttributeDescriptorViewModel`, and the `ResultLimit` with the `TokenizedTopicListAttributeDescriptorViewModel`. In addition, it currently indirectly sets the `AutoPostBack`, `AttributeKey`, and `AttributeValue`.

By deriving directly from it, not only do we reduce redundancy, but we provide more flexibility for cases where this is being used for general topic references, and not the `BaseTopic` specifically. E.g., we allow topic references to toggle `AutoPostBack` or set alternate attribute filters.

This begins to address the needs of #27.