Customize view model in `TopicReferenceViewComponent`
Now that the `TopicReferenceAttributeDescriptorViewModel` derives from the `TokenizedTopicListAttributeDescriptorViewModel`, we can just set the properties we want to customize directly on the `TopicReferenceAttributeDescriptorViewModel` and then relay that to the underlying `TokenizedTopicListViewComponent`.

This has a few benefits. First, it's cleaner. Second, it eliminates the need to define interstitial properties on the `TopicReferenceAttributeDescriptorViewModel` exclusively for configuring the `TokenizedTopicListViewComponent`. Third, it makes the logic clearer in the view component itself.

As part of this, I limit the customizations to attributes named `BaseTopic`. This isn't ideal. But it maintains backward compatibility with OnTopic 5.0.0, which used the `ReferenceTopicAttribute` exclusively for `BaseTopic`, without introducing the need to add new attributes in order to maintain backward compatibility. We may further abstract this by introducing a special attribute later, but as requiring this would be a breaking change, this hard-coded value won't be removed until OnTopic 6.0.0.