Introduced support for `UseCurrentContentType`
The `UseCurrentContentType` is intended to abstract the behavior required by `BaseTopic`, where results should be filtered by the current content type. In the future, we'll remove the hard-coded check for `BaseTopic` and rely exclusively on `UseCurrentContentType`. For OnTopic 5.x, however, the `BaseTopic` reference is needed for backward compatibility, since the original OnTopic 5.0.0 release didn't expect `UseCurrentContentType` (7bba5f6).

This satisfies the expectations of #27.