Updated JavaScript to target buttons for error validation
In a previous commit, we updated the markup for the tabs to use buttons instead of anchor links, in order to remain consistent with Bootstrap's reference examples in their documentation (258d0b1). This broke the JavaScript selectors for highlighting tabs with errors on them, since it was expecting an anchor link.