Consolidated scripts for initializing Bootstrap tooltips
In a previous commit, I established a new script for initializing the Bootstrap tooltips (9b5ea80)—while overlooking that we already had one in place. Whoops! This consolidates this code into `UserInterface.js` (where it makes more sense anyway) and merges it with the previous configuration that it was using.