Introduced new `IsProtected` property
The `IsProtected` property communicates to the OnTopic Editor that a `Topic` should not be able to be deleted or  moved. This is similar to the `DisableDelete`, except it can be applied to any topic, not just `ContentTypeDescriptor`s, and it applies to the immediate topic, not topics of that content type. This is the foundation for #35.