Renamed `isDraggable` to `isMovable`
ExtJS looks for the `draggable` attribute to determine if a node can be dragged. By changing this to `isMovable`, we prevent this functionality. This, in turn, allows us to access this property to validate whether a topic can be moved to a specific location. This is in service of allowing `IsProtected` and `DisableDelete` topics to be _reordered_, but not _moved_, as per #45.