Update validation logic to utilize `isMovable`
Instead of validating the `startdrag` event, instead validate the `nodedragover` event which allows us to evaluate both the `dropNode` source and the `target` node. The logic is updated to allow reordering, but not moves, by evaluating both `isMovable` (5a1fcf6) as well as the path of the parent nodes. This completes the implementation of #45.