Introduced `ModelType.Reflexive` support into `EditorController`
In preparation for establishing a new `ReflexiveAttribute` (#44) , and with the `ReflexiveAttributeDescriptor` established (10788b7), the `EditorController` now has special handling for `ModelType.Reflexive`. If this is set, it will instead set the `ModelType` to that of the `CurrentTopic`, which is assumed to be an `AttributeDescriptor`. This allows the value for an attribute marked as `ModelType.Reflexive` to be pulled and stored from the same storage location as the `AttributeDescriptor` it's meant to reflect. This is important as we don't want e.g. topic references storing the `topic.Id` in `Topic.Attributes`, where referential integrity isn't maintained.