Introduced `ReflexiveViewComponent`
The `ReflexiveViewComponent` is at the heart of the `ReflexiveAttribute` implementation (#44). It takes an `ITopicMappingService` and an `ITopicRepository` so that it can remap the `currentTopic` as whatever `AttributeDescriptorViewModel` it should be associated with, modifies the core `AttributeDescriptor` properties so that it has the `Key`, `Title`, `Description`, &c. of the `ReflexiveAttribute`, and then returns that type as the view model. This will allow that view model to be used to invoke whichever `AttributeDescriptor` the `ReflexiveAttribute` is reflecting, while still being located and labeled as expected by the `ReflexiveAttribute` configuration.

The `ITypeLookupService` and `ITopicMappingService` are unfortunate artifacts, which stem from the fact that while we have the `currentTopic` mapped, it is mapped as an `EditingTopicViewModel`, and not a derivative of the `AttributeDescriptorTopicViewModel`, and thus won't have the properties available that the target attribute type view component will be expecting. As such, we need to look it up, remap it according to its `ContentType`, and modify that view model.

There might be a way to handle this via the `EditorController`, but it's challenging since `ReflexiveAttributeDescriptorViewModel` doesn't derive from the `EditingTopicViewModel`—and we don't want it to, since `EditingTopicViewModel` comes with a lot of extra mapped properties that aren't generally needed _outside_ of cases where a `ReflexiveAttribute` is present. Similarly, we don't want to map a duplicate copy as an `AttributeDescriptorViewModel` just in case it's needed by a `ReflexiveAttribute`. As such, we're maintaining these dependencies so that the `ReflexiveViewComponent` can map its own dependencies.