Wire up dependencies in `StandardEditorComposer`
Unfortunately, the `ITypeLookupService` and `ITopicMappingService` dependencies are new, and not currently requested by the `StandardEditorComposer`. Introducing them as dependencies would be a breaking change, however, and would thus need to wait until OnTopic 6.0.0.

Fortunately, we know we'll _always_ want the built-in, well-known `EditorViewModelLookupService`, so maintaining a dependency on that isn't a big deal—in fact, we take a similar approach with the `AttributeBindingModelLookupService`, which is hard-coded into the `AttributeBindingModelBinder`.

Further, it's highly unlikely we'd ever expect a different `TopicMappingService` than the well-known implementation in the `OnTopic` library—which is an established dependency of the OnTopic Editor regardless. And, if we did, implementers can always wire up the `ReflexiveViewComponent` directly, if needed, instead of relying on the `StandardEditorComposer`.

Still, this isn't ideal from a dependency injection perspective, as it means underlying dependencies are not clearly communicated via the constructor. Given that, we'll likely want to introduce these to the constructor in OnTopic Editor 6.0.0.