Introduced the `ReflexiveAttribute` view
The `ReflexiveAttribute` view is pretty straight forward: It simply invokes the attribute type view component associated with the `CurrentTopic` (which we assume to be an `AttributeDescriptor`) If the `CurrentTopic` is _not_ an `AttributeDescriptor`, or it is the `ReflexiveAttribute` itself, a warning is displayed. Otherwise, the `CurrentTopic` is rendered as a view component, thus reflecting itself, and fulfilling the requirements of #44.