Excluded relationships and nested topics from `ReflexiveAttribute`
In general, the use cases for the `ReflexiveAttribute`—such as a `DefaultValue` or an `ImplicitValue`—don't really make sense with relationships or nested topics. We may revisit this later, but for now we're excluding `AttributeDescriptorViewModel`s with a `ModelType` of either `Relationship` or `NestedTopic`.

As part of this, I refactored the `isReflexive` variable to a new `isSupported` variable, which relies on the `Model.AttributeDescriptor.ModelType` instead of the `Model.CurrentTopic.ContentType`. This effectively accomplishes the same thing, but allows isn't specific to any one attribute type, instead factoring in how the data is stored.

This finalize the implementation of #44.