Updated `AttributeBindingModel.Value` to allow `null` values
In a previous update, I switched from `[NotNull, DisallowNull]` to `= default!`. These _should_ be syntactically equivalent. But they're not actually treated that way by the model binding validator, which will allow `[DisallowNull]` on a nullable type, but not on a non-nullable type. This is an issue because radio buttons, in particular, might be null—which is why we don't validate `Value` in the `AttributeBindingModelBinder`.

This can be fixed by making `Value` nullable. This has some downstream effects, which will be addressed separately.