Account for potentially nullable `AttributeBindingModel.Value`
With the `AttributeBindingModel.Value` now allowing null values (63e8b40), we need to account for that to prevent a null reference exception. With the `Key` setting, we assume a `Value` exists, as `Key` should always be `IsRequired`.

With relationships, we do the same thing, as while relationships aren't required, we expect an empty value if no relationships are being submitted; if that isn't true, that's a sign of a model binding issue, and should result in an exception.