Established `RegisterResources()` virtual method
In preparation for the introduction of a new constructor, established a `RegisterResources` method for the models that need to register stylesheets or scripts, as to centralize these registrations between multiple constructors.

I had considered establishing this as a virtual method off of `AttributeDescriptorViewModel` and calling it from the base constructor. Unfortunately, this throws the CA2214 warning, which relates to the overridden methods that rely on state established in a derived constructor will fail, since the base constructor gets called before the derived constructor. That isn't an _expected_ use case here, but it's certainly one we can _imagine_; e.g., an attribute exposing a `Theme` property for determining which style sheet associated with a plugin to load.

Given that, I'm instead calling this directly from the local constructors.