Upgrade to the OnTopic 5.2.0 preview release
This version of the library isn't complete, and is only available on Ignia's internal NuGet feed, but allows us to begin relying on OnTopic 5.2.0 specific features in the OnTopic Editor—most notably, the new `AttributeDictionary` constructors, which promise to be faster when mapping models with a lot of scalar values.

This version will need to be updated to the final 5.2.0 release prior to release. Further, it is anticipated that the name of the `AttributeDictionary` library may change prior to then.