Established `AttributeDictionary` constructor overload
The `AttributeDictionary` provides a faster way to establish local property values during the mapping process than relying on reflection. When a constructor accepting an `AttributeDictionary` as its only parameter is exposed, the `TopicMappingLibrary` will pass a lightweight `AttributeDictionary` containing key/value pairs of all attributes from the current topic and any `BaseTopic` references. This puts the onus on the constructor to set these values, and then bypasses subsequent processing of them in the `TopicMappingLibrary`. When there are multiple mapped attributes, this proves to be significantly faster than relying on reflection to set each property value. This is a new capability of OnTopic 5.2.0.

This addresses the core requirement of Feature #49.