Introduced `PermittedContentTypeViewModel` for content type references
Every `ContentTypeDescriptorViewModel` has a `PermittedContentTypes` property that references other `ContentTypeDescriptor`s that are permitted to be created under it. These are exclusively needed for the `ContentTypeList` view component, which only needs the `Key` and `Title` property. And, yet, currently, the entire `ContentTypeDescriptorViewModel` is mapped, including those derived from the underlying `EditingTopicViewModel` and `CoreTopicViewModel` (excluding associations). That's a lot of overhead for every request just to render a dropdown menu.

To mitigate this, the new `PermittedContentTypeViewModel` is introduced for mapping _references_ to content type descriptors without actually mapping the entire `ContentTypeDescriptor`. This is implemented in the `ContentTypeDescriptorViewModel` (obviously), but also the components that rely on it, such as the `ContentTypeList` view component and its corresponding view model.

This addresses the requirements for Improvement #50.